chore: move create blog posts into own file
